+++ Classes/GUI/QuickLookTableView.m
@@ -24,83 +24,66 @@
     }
 }
 
-- (void)highlightSelectionInClipRect:(NSRect)clipRect
+
+// Override The row colour with the label colour for only the filename column.
+// with nice rounded side if there is a label colour.
+- (void)drawRow:(NSInteger)row clipRect:(NSRect)clipRect;
 {
 	NSIndexSet *selectedRowIndexes = [self selectedRowIndexes];
-	NSRange visibleRows = [self rowsInRect:clipRect];
-	
-	NSUInteger selectedRow = [selectedRowIndexes firstIndex];
-    while (selectedRow != NSNotFound)
-    {
-		if (selectedRow == -1 || !NSLocationInRange(selectedRow, visibleRows)) {
-			selectedRow = [selectedRowIndexes indexGreaterThanIndex:selectedRow];
-			continue;
-		}		
-		
-		if ( [[self window] isKeyWindow] ) {
-			[[NSColor alternateSelectedControlColor] set];
-		}
-		else {
-			[[NSColor secondarySelectedControlColor] set];
-		}
-		
-		NSRect rowRect = [self rectOfRow:selectedRow];
-		NSRectFill(rowRect);
-		
+	if ([selectedRowIndexes containsIndex:row]){
+				
 		// draw a "dot" of color, similar to what the Finder does, 
 		NSInteger labelIndex = BHLabelColorNone;
 		if ([[self dataSource] respondsToSelector:@selector(labelColorForRow:)]) {
-			labelIndex = [(id)[self dataSource] labelColorForRow:selectedRow];
+			labelIndex = [(id)[self dataSource] labelColorForRow:row];
 		}
 		NSGradient *gradient = [BHFinderLabelColours gradientForLabelIndex:labelIndex];
 		
 		if (gradient)
 		{			
+			NSRect rowRect = [self rectOfRow:row];
+			NSRect colRect = [self rectOfColumn: [self columnWithIdentifier:@"filename"]];
+			rowRect.origin.x   = colRect.origin.x;
+			rowRect.size.width = colRect.size.width;
 			NSRect gradientRect = 
 			NSMakeRect(2, 
-					   ([self rowHeight]+[self intercellSpacing].height)*selectedRow + 1.0, 
+					   ([self rowHeight]+[self intercellSpacing].height)*row + 1.0, 
 					   20.0, 
 					   rowRect.size.height - 2.0);
 			NSBezierPath *gradientPath = [NSBezierPath bezierPathWithRoundedRect:gradientRect 
 																		 xRadius:9.0 
 																		 yRadius:9.0];
 			[gradient drawInBezierPath:gradientPath angle:90.0];			
+		}		
+		[super drawRow:row clipRect:clipRect];	
+	}else{
+		NSInteger labelIndex= BHLabelColorNone;
+		if ([[self dataSource] respondsToSelector:@selector(labelColorForRow:)]) {
+			labelIndex = [(id)[self dataSource] labelColorForRow:row];
 		}
 		
-		selectedRow = [selectedRowIndexes indexGreaterThanIndex:selectedRow];
-    }
-}
-
-// Override The row colour with the label colour for only the filename column.
-// with nice rounded side if there is a label colour.
-- (void)drawRow:(NSInteger)row clipRect:(NSRect)clipRect;
-{
-	NSInteger labelIndex= BHLabelColorNone;
-	if ([[self dataSource] respondsToSelector:@selector(labelColorForRow:)]) {
-		labelIndex = [(id)[self dataSource] labelColorForRow:row];
+		NSGradient *gradient = [BHFinderLabelColours gradientForLabelIndex:labelIndex];
+		
+		if (gradient && ![[self selectedRowIndexes] containsIndex:row])
+		{
+			NSRect rowRect = [self rectOfRow:row];
+			NSRect colRect = [self rectOfColumn: [self columnWithIdentifier:@"filename"]];
+			rowRect.origin.x   = colRect.origin.x;
+			rowRect.size.width = colRect.size.width;
+			
+			NSRect gradientRect = NSMakeRect(0, 
+											 ([self rowHeight]+[self intercellSpacing].height)*row + 1.0, 
+											 rowRect.size.width - 4.0, 
+											 rowRect.size.height - 2.0);
+			NSBezierPath *smallerPath = [NSBezierPath bezierPathWithRoundedRect:gradientRect 
+																		xRadius:9.0 
+																		yRadius:9.0];
+			[gradient drawInBezierPath:smallerPath 
+								 angle:90.0];
+		}
+		
+		[super drawRow:row clipRect:clipRect];	
 	}
-	
-	NSGradient *gradient = [BHFinderLabelColours gradientForLabelIndex:labelIndex];
-	
-	if (gradient && ![[self selectedRowIndexes] containsIndex:row])
-	{
-		NSRect rowRect = [self rectOfRow:row];
-		NSRect colRect = [self rectOfColumn: [self columnWithIdentifier:@"filename"]];
-		rowRect.origin.x   = colRect.origin.x;
-		rowRect.size.width = colRect.size.width;
-
-		NSRect gradientRect = NSMakeRect(0, 
-										 ([self rowHeight]+[self intercellSpacing].height)*row + 1.0, 
-										 rowRect.size.width - 4.0, 
-										 rowRect.size.height - 2.0);
-		NSBezierPath *smallerPath = [NSBezierPath bezierPathWithRoundedRect:gradientRect 
-																	xRadius:9.0 
-																	yRadius:9.0];
-		[gradient drawInBezierPath:smallerPath 
-							 angle:90.0];
-	}
-	
-	[super drawRow:row clipRect:clipRect];
 }